pyproject.toml
[project]
name = "ludii-automation"
version = "0.1.0"
description = "Automation scripts including mouse tracking and OCR features."
requires-python = ">=3.9"
dependencies = [
    "pyautogui>=0.9.54",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
